Refactored convert helpers
