+++ pkg/apis/apiserver/convert/organization_unstructured_conversion.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apiserver
+package convert
 
 import (
 	"fmt"
@@ -23,6 +23,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 
+	"github.com/kubermatic/bulward/pkg/apis/apiserver"
 	corev1alpha1 "github.com/kubermatic/bulward/pkg/apis/core/v1alpha1"
 )
 
@@ -36,7 +37,7 @@ func chainConversion(scheme *runtime.Scheme, initObj runtime.Object, objs ...run
 	return objs[len(objs)-1], nil
 }
 
-func ConvertToV1Alpha1Unstructured(organization *Organization, scheme *runtime.Scheme) (*unstructured.Unstructured, error) {
+func ToUnstructuredCoreV1Alpha1Organization(organization *apiserver.Organization, scheme *runtime.Scheme) (*unstructured.Unstructured, error) {
 	u := &unstructured.Unstructured{}
 	if _, err := chainConversion(scheme, organization, &corev1alpha1.Organization{}, u); err != nil {
 		return nil, err
@@ -44,7 +45,7 @@ func ConvertToV1Alpha1Unstructured(organization *Organization, scheme *runtime.S
 	return u, nil
 }
 
-func ConvertFromV1Alpha1Unstructured(internalOrgv1alpha1 *unstructured.Unstructured, scheme *runtime.Scheme) (*Organization, error) {
+func FromUnstructuredCoreV1Alpha1(internalOrgv1alpha1 *unstructured.Unstructured, scheme *runtime.Scheme) (*apiserver.Organization, error) {
 	gvk, err := apiutil.GVKForObject(internalOrgv1alpha1, scheme)
 	if err != nil {
 		return nil, err
@@ -53,17 +54,17 @@ func ConvertFromV1Alpha1Unstructured(internalOrgv1alpha1 *unstructured.Unstructu
 	if gvk != expectedGVK {
 		return nil, fmt.Errorf("wrong GVK, expected %v, found %v", expectedGVK, gvk)
 	}
-	org := &Organization{}
+	org := &apiserver.Organization{}
 	if _, err := chainConversion(scheme, internalOrgv1alpha1, &corev1alpha1.Organization{}, org); err != nil {
 		return nil, err
 	}
 	return org, nil
 }
 
-func ConvertFromV1Alpha1UnstructuredList(internalOrgv1alpha1 *unstructured.UnstructuredList, scheme *runtime.Scheme) (*OrganizationList, error) {
-	sol := &OrganizationList{}
-	for _, it := range internalOrgv1alpha1.Items {
-		org, err := ConvertFromV1Alpha1Unstructured(&it, scheme)
+func ToUnstructuredCoreV1Alpha1OrganizationList(organizations *apiserver.OrganizationList, scheme *runtime.Scheme) (*unstructured.UnstructuredList, error) {
+	sol := &unstructured.UnstructuredList{}
+	for _, it := range organizations.Items {
+		org, err := ToUnstructuredCoreV1Alpha1Organization(&it, scheme)
 		if err != nil {
 			return nil, err
 		}
@@ -72,10 +73,10 @@ func ConvertFromV1Alpha1UnstructuredList(internalOrgv1alpha1 *unstructured.Unstr
 	return sol, nil
 }
 
-func ConvertToV1Alpha1UnstructuredList(organizations *OrganizationList, scheme *runtime.Scheme) (*unstructured.UnstructuredList, error) {
-	sol := &unstructured.UnstructuredList{}
-	for _, it := range organizations.Items {
-		org, err := ConvertToV1Alpha1Unstructured(&it, scheme)
+func FromUnstructuredCoreV1Alpha1List(internalOrgv1alpha1 *unstructured.UnstructuredList, scheme *runtime.Scheme) (*apiserver.OrganizationList, error) {
+	sol := &apiserver.OrganizationList{}
+	for _, it := range internalOrgv1alpha1.Items {
+		org, err := FromUnstructuredCoreV1Alpha1(&it, scheme)
 		if err != nil {
 			return nil, err
 		}
